refactor: ReservationId 로 리턴값 추상화
