--- src/main/java/roomescape/domain/ReservationId.java
@@ -1,28 +1,0 @@
-package roomescape.domain;
-
-import java.util.Objects;
-
-public class ReservationId {
-    private final long id;
-
-    public ReservationId(long id) {
-        this.id = id;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        ReservationId that = (ReservationId) o;
-        return id == that.id;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
