+++ src/main/java/roomescape/dto/ReservationId.java
@@ -1,0 +1,4 @@
+package roomescape.dto;
+
+public record ReservationId(long id) {
+}
